requirements.txt
Babel==2.6.0
Django==2.0.7
django-collectionfield==0.0.4
django-datetime-widget==0.9.3
django-datetimepicker==3.14
django-decorator-include==2.1
django-mathfilters==0.4.0
django-phonenumber-field==2.1.0
django-sortedm2m==1.5.0
django-sortedm2m-filter-horizontal-widget==1.3.2
django-updateable==0.2.1
django-widget-tweaks==1.4.3
mysqlclient==1.4.2.post1
phonenumbers==8.10.2
Pillow==5.3.0
pkg-resources==0.0.0
psycopg2-binary==2.8.2
python-dateutil==2.7.5
pytz==2018.7
six==1.12.0
sqlparse==0.3.0
